@@ -1,18 +1,14 @@
-import React, { createContext, useContext, useEffect, useState } from "react";
-
+import React, { createContext, useState, useEffect, useContext } from "react";
 const Crypto = createContext();
-
 const CryptoContext = ({ children }) => {
   const [currency, setCurrency] = useState("INR");
   const [symbol, setSymbol] = useState("₹");
-
   useEffect(() => {
     if (currency === "INR") setSymbol("₹");
     else if (currency === "USD") setSymbol("$");
   }, [currency]);
-
   return (
-    <Crypto.Provider value={{ currency, setCurrency, symbol }}>
+    <Crypto.Provider value={{ currency, symbol, setCurrency, setSymbol }}>
       {children}
     </Crypto.Provider>
   );
